fix: 회의에서 이야기한대로 수정
