+++ just-swim-app/src/app/(afterLogin)/instructor/_component/BottomNavBar.tsx
@@ -1,0 +1,42 @@
+"use client"
+
+import Link from 'next/link';
+import styles from './BottomNavBar.module.css';
+import { useState } from 'react';
+
+export default function BottomNavBar() {
+  const [type, setType] = useState('schedule');
+
+  const handleBottomNavClick = (pageType: string) => {
+    setType(pageType);
+  };
+
+  return (
+    <div className={styles.footer}>
+      {/* 네비게이션 바 */}
+      <div className={styles.bottomNavBarWrapper}>
+        <Link href="/instructor/classManagement">
+          <button
+            className={`${type === 'classManagement' ? styles.isActive : ''}`}
+            onClick={() => handleBottomNavClick('classManagement')}>
+            수업
+          </button>
+        </Link>
+        <Link href="/instructor/schedule">
+          <button
+            className={`${type === 'schedule' ? styles.isActive : ''}`}
+            onClick={() => handleBottomNavClick('schedule')}>
+            홈
+          </button>
+        </Link>
+        <Link href="/instructor/feedback">
+          <button
+            className={`${type === 'feedback' ? styles.isActive : ''}`}
+            onClick={() => handleBottomNavClick('feedback')}>
+            피드백
+          </button>
+        </Link>
+      </div>
+    </div>
+  );
+}
